add profesor

+++ src/main/java/ro/upb/smartfeedback/repository/ProfesorRepository.java
@@ -8,6 +8,6 @@ import ro.upb.smartfeedback.entity.Profesor;
  * Created by colez on 29/11/2015.
  */
 @Repository
-public interface ProfesorRepository extends JpaRepository<Profesor, Integer>{
-    public Profesor getById(Integer id);
+public interface ProfesorRepository extends JpaRepository<Profesor, Long>{
+    public Profesor getById(Long id);
 }
